backend/requirements.txt
# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23

# Authentication and security
PyJWT==2.8.0
python-dotenv==1.0.0

# Data validation and models
pydantic==2.5.0

# AI and machine learning
google-generativeai==0.3.2

# Data processing and analysis
pandas==2.1.4
openpyxl==3.1.2

# Additional utilities 